Caching improvements
Before the plugin checked if any file in your mods were changed and if it was every object was recreated, taking a lot more time than needed. Now it checks if any of the subfolders in your mod that game objects are read from have changed and keeps a hash for each subfolder that is relevant to the plugin and then only creates the objects that actually have been changed. This significantly increases startup performance and object/cache creation.